% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provExplainR.R
\name{print.input.files.changes}
\alias{print.input.files.changes}
\title{print.input.files.changes checks if input files in two data provenance
collections are different based on their hash values.
NOTE: this function currently does not work for URL as input.}
\usage{
\method{print}{input.files.changes}(input.df1, input.df2)
}
\arguments{
\item{input.df1}{data frame of input files in first data provenance collection}

\item{input.df2}{data frame of input files in second data provenance collection}
}
\description{
print.input.files.changes checks if input files in two data provenance
collections are different based on their hash values.
NOTE: this function currently does not work for URL as input.
}
