% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provExplainR.R
\name{prov.diff.script}
\alias{prov.diff.script}
\title{Provenance Script Diff function}
\usage{
prov.diff.script(first.script, dir1, dir2, second.script = NULL)
}
\arguments{
\item{first.script}{name of first script}

\item{dir1}{path of first provenance directory}

\item{dir2}{path of second provenance directory}

\item{second.script}{name of second script}
}
\description{
prov.diff.script visualizes the differences between the contents of two scripts based on given provenance collections.
}
\details{
This function uses provenance collected using the rdtLite or rdt packages.

Users must specify name of the first script, 
first provenance directory path and second provenance directory 
path. Name of second script is optional. If second script 
is specified, provExplainR assumes first script is located in the 
first provenance directory and second script is located in 
the second provenance directory. Otherwise, provExplainR assumes 
both scripts share the same name.
}
