% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provExplainR.R
\name{prov.diff.script}
\alias{prov.diff.script}
\title{Provenance Script Diff function}
\usage{
prov.diff.script(first.script, olderProv.dir, newerProv.dir,
  second.script = NULL)
}
\arguments{
\item{first.script}{name of first script}

\item{olderProv.dir}{path of older provenance directory}

\item{newerProv.dir}{path of newer provenance directory}

\item{second.script}{name of second script}
}
\description{
prov.diff.script visualizes the differences between the contents of two scripts based on given provenance collections.
}
\details{
This function use provenance collected using the rdtLite or rdt packages.

Users must specify name of the first script, 
old provenance directory path and new provenance directory 
path. Name of second script is optional. If second script 
is specified, provExplainR assumes first script is located in the 
old provenance folder and second script is located in 
the new provenance folder. Otherwise, provExplainR assumes 
both scripts share the same name.
}
