% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provExplainR.R
\name{prov.explain}
\alias{prov.explain}
\title{Immplementation of provExplainR: detect changes and report to users}
\usage{
prov.explain(olderProv.dir, newerProv.dir, save = FALSE)
}
\arguments{
\item{olderProv.dir}{path of older provenance directory}

\item{newerProv.dir}{path of newer provenance directory}

\item{save}{if true saves the report to the file prov-explain.txt
in the current working directory}
}
\description{
Immplementation of provExplainR: detect changes and report to users
}
\author{
Khanh Ngo
provExplainR takes in two provenance directories, detect changes, and
report to users. Set 'save' parameter to be true to have the results
written in a txt file.
}
